src/main/java/com/dgalicki/room/occupancy/service/CustomerServiceImpl.java
package com.dgalicki.room.occupancy.service;

import com.dgalicki.room.occupancy.model.entity.Customer;
import com.dgalicki.room.occupancy.repository.CustomerRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class CustomerServiceImpl implements CustomerService {

    private final CustomerRepository customerRepository;

    @Override
    public Customer saveCustomer(Customer customer) {
        return customerRepository.save(customer);
    }
}